requirements.txt
Flask
pandas
requests
mysql-connector-python
